pyproject.toml
[project]
name = "my-sqlalchemy-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "isort>=6.0.1",
    "pytest>=8.3.5",
    "sqlalchemy>=2.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "ipykernel>=6.29.5",
]

[tool.setuptools]
packages = ["app"]


[tool.uv]
package = true

[tool.ruff]
line-length = 120
